ui-服务详情，导游详情，导游列表，订购页，支付页，评价列表等

+++ LetsGoApp/app/src/main/java/com/njz/letsgoapp/adapter/home/GuideListAdapter.java
@@ -57,7 +57,7 @@ public class GuideListAdapter extends RecyclerView.Adapter<RecyclerView.ViewHold
             ((GuideViewHolder) holder).rating_bar_route.setRating((int)data.getGuideScore());
             ((GuideViewHolder) holder).tv_content.setText(data.getIntroduce());
             ((GuideViewHolder) holder).tv_service_item.setServiceTag(data.getServiceTags());
-            ((GuideViewHolder) holder).ll_times.setGuideScore(data.getServiceCounts(),(int)data.getGuideScore(),data.getCount());
+            ((GuideViewHolder) holder).ll_times.setGuideScore(data.getServiceCounts(),data.getGuideScore(),data.getCount());
             if (mOnItemClickListener != null) {
                 ((GuideViewHolder) holder).cardView.setOnClickListener(new View.OnClickListener() {
                     @Override
